refactor(hub): rename generic extension modules (#444)
