+++ backend/app/api/routers/__init__.py
@@ -11,8 +11,8 @@ ROUTER_MODULES: Final[tuple[str, ...]] = (
     "app.api.routers.admin_a2a_agents",
     "app.api.routers.admin_proxy_allowlist",
     "app.api.routers.a2a_schedules",
-    "app.api.routers.a2a_extensions_opencode",
-    "app.api.routers.hub_a2a_extensions_opencode",
+    "app.api.routers.a2a_extension_capabilities",
+    "app.api.routers.hub_a2a_extension_capabilities",
     "app.api.routers.opencode_session_directory",
     "app.api.routers.me_sessions",
     "app.api.routers.invitations",
