+++ backend/app/api/routers/_extension_capability_router.py
