+++ backend/tests/test_interrupt_callback_extension_discovery.py
