+++ backend/tests/test_session_query_extension_discovery.py
